+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Add login word variant for Spanish.
 
+## Changed
+- Search also for login elements with ARIA role button.
 
 ## [0.28.0] - 2025-09-02
 ### Added